requirements.txt
black==20.8b1
psycopg2-binary
psutil==5.8.0